+++ nc_py_api/__init__.py
@@ -3,7 +3,7 @@ Import all possible stuff that can be used.
 """
 
 from ._version import __version__
-from .constants import ApiScope, LogLvl
+from .constants import ApiScope, LogLvl, SharePermissions, ShareStatus, ShareType
 from .exceptions import NextcloudException, NextcloudExceptionNotFound, check_error
 from .files import FsNode, FsNodeInfo
 from .integration_fastapi import (