+++ nc_py_api/constants.py
@@ -36,21 +36,58 @@ class OCSRespond(IntEnum):
     RESPOND_UNKNOWN_ERROR = 999
 
 
+class SharePermissions(IntEnum):
+    """The share permissions to be set.
+
+    All permissions can be combined with each other, except ``PERMISSION_ALL``"""
+
+    PERMISSION_READ = 1
+    """Access to read"""
+    PERMISSION_UPDATE = 2
+    """Access to write"""
+    PERMISSION_CREATE = 4
+    """Access to create new objects in the share"""
+    PERMISSION_DELETE = 8
+    """Access to remove objects in the share"""
+    PERMISSION_SHARE = 16
+    """Access to re-share objects in the share"""
+    PERMISSION_ALL = 31
+    """Full access to the shared object"""
+
+
 class ShareType(IntEnum):
+    """Type of the object that will receive share"""
+
     TYPE_USER = 0
+    """Share to the user"""
     TYPE_GROUP = 1
+    """Share to the group"""
     TYPE_LINK = 3
+    """Share by link"""
     TYPE_EMAIL = 4
+    """Share by the email"""
     TYPE_REMOTE = 6
+    """Share to the Federation"""
     TYPE_CIRCLE = 7
+    """Share to the Nextcloud Circle"""
     TYPE_GUEST = 8
+    """Share to `Guest`"""
     TYPE_REMOTE_GROUP = 9
+    """Share to the Federation group"""
     TYPE_ROOM = 10
+    """Share to the Talk room"""
     TYPE_DECK = 11
-    TYPE_SCIENCEMESH = 15
+    """Share to the Nextcloud Deck"""
+    TYPE_SCIENCE_MESH = 15
+    """Share to the Reva instance(Science Mesh)"""
 
 
 class ShareStatus(IntEnum):
+    """Status of the share"""
+
     STATUS_PENDING = 0
+    """The share waits for acceptance"""
     STATUS_ACCEPTED = 1
+    """The share was for accepted"""
     STATUS_REJECTED = 2
+    """The share was for rejected"""